Restructure highlighting by turning into modules
- Add a modules/ directory in src/ to hold re-usable ES modules
- Create a common module with functions and types used across the whole
  content script
- Create a highlight/ directory in src/modules/ to hold all
  highlighting-related modules
- Split off all highlighting components into modules importing from each
  other
- Do some of the work towards making imports work in the browser
  extension environment
  - Add a script to automatically convert TypeScript-friendly imports
    into browser-friendly imports
- Edit manifest to expose module scripts
