+++ src/modules/highlight/highlighting.mts
@@ -1,0 +1,21 @@
+const { getElementTagsSet } = await import("src/modules/common.mjs");
+
+const HIGHLIGHT_TAG = "mms-h";
+const HIGHLIGHT_TAG_UPPER = "MMS-H";
+
+type HighlightTags = {
+	reject: ReadonlySet<string>,
+	flow: ReadonlySet<string>,
+}
+
+const highlightTags: HighlightTags = {
+	reject: getElementTagsSet([ "meta", "style", "script", "noscript", "title", "textarea" ]),
+	flow: getElementTagsSet([ "b", "i", "u", "strong", "em", "cite", "span", "mark", "wbr", "code", "data", "dfn", "ins",
+		HIGHLIGHT_TAG as keyof HTMLElementTagNameMap ]),
+	// break: any other class of element
+};
+
+export {
+	HighlightTags, highlightTags,
+	HIGHLIGHT_TAG, HIGHLIGHT_TAG_UPPER,
+}
