+++ tsconfig.json
@@ -2,7 +2,12 @@
 	"compilerOptions": {
 		"outDir": "dist",
 		"target": "ES2022",
+		"module": "ES2022",
+		"declaration": true,
 		"alwaysStrict": true,
 		"strictNullChecks": true,
 	},
+	"include": [
+		"src/**/*",
+	],
 }
